splitting up block.rs
